requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10

# Ollama integration
ollama==0.1.6

# Vector stores
chromadb==0.4.22
faiss-cpu==1.7.4

# Document loaders
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.2

# Text processing
sentence-transformers==2.2.2

# Web UI
streamlit==1.29.0
gradio==4.13.0

# CLI utilities
click==8.1.7
rich==13.7.0

# Environment management
python-dotenv==1.0.0
